LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/embed.https.sub-expected.txt



PASS Wrapper: same-origin embed
PASS Wrapper: Navigate to same-origin embed
FAIL Wrapper: same-site embed assert_not_equals: got disallowed value "No header has been recorded"
FAIL Wrapper: Navigate to same-site embed assert_not_equals: got disallowed value "No header has been recorded"
FAIL Wrapper: cross-site embed assert_not_equals: got disallowed value "No header has been recorded"
FAIL Wrapper: Navigate to cross-site embed assert_not_equals: got disallowed value "No header has been recorded"
FAIL same-origin embed: sec-fetch-dest assert_equals: expected "embed" but got ""
FAIL same-origin embed: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL same-origin embed: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS same-origin embed: sec-fetch-user
FAIL Navigate to same-origin embed: sec-fetch-dest assert_equals: expected "embed" but got ""
FAIL Navigate to same-origin embed: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL Navigate to same-origin embed: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS Navigate to same-origin embed: sec-fetch-user
